add xsql-test crate for dev-dep.
