+++ core/rpc/Cargo.toml
@@ -43,3 +43,4 @@ rand = "0.8"
 smallvec = "1.6"
 
 core-cli = { path = "../cli" }
+xsql-test = { path = "../../db/xsql-test" }
