+++ core/rpc/src/tests/mod.rs
@@ -90,7 +90,9 @@ impl RpcTestEngine {
             .connect(DBDriver::SQLite, MEMORY_DB, "", 0, "", "")
             .await
             .unwrap();
-        store.create_tables().await.unwrap();
+
+        let mut tx = store.inner.transaction().await.unwrap();
+        xsql_test::create_tables(&mut tx).await.unwrap();
 
         let config: MercuryConfig = parse(CONFIG_PATH).unwrap();
         let script_map = config.to_script_map();
