+++ core/storage/Cargo.toml
@@ -13,4 +13,3 @@ ckb-types = "0.43"
 common = { path = "../../common" }
 protocol = { path = "../../db/protocol" }
 xsql = { path = "../../db/xsql" }
-
