+++ db/xsql-test/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "xsql-test"
+version = "0.1.0"
+authors = ["EthanYuan <yuanchengxing@hotmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rbatis = { git = "https://github.com/KaoImin/rbatis.git", branch = "bson", features = ["all-database", "tokio1"] }
+
+common = { path = "../../common" }
